Usable OMP threading for long_range (core-level).
The ideal level of threading evenly divides the number of L/spin channels.
The is ( 2 * l_c + 1 ) * 2 * 2. Where l_c is the angular quantum number of
the core level, one factor of 2 from the core spin, and one from the valence.
Best is num_threads = 1, 2, or 4 (K edges)  = 1, 2, 3, 4, 6, 12 (L edges).