Fixes pawsetup.x for multiple 'types' of same 'Z' #205
This is a problem with QE runs such as NiO where the Ni sites must be labeled
differently for spin up and spin down. Or for a more complicated system where
a transition metal might have different U values.

By tracking site index by Z instead of by typat we can now count correctly.

This will still not work if the user wants to use different pseudopotentials.
I'm not sure if that is a reasonable case, but will require restructuring
OPF and CNBSE directories along with the SCREEN directory.