Adding 'allow' to core-level
The allow vector applies the Fermi function/occupation numbers to the
particle-hole basis. This has always been there for the valence code
to work with metals, but is new for the core-level code. This will make
things easier for metals and spin systems, espcially for integrating with
the valence code, i.e., RIXS.