Bugs associated with screening of nspin=2
1. Minor change to vcbder2 which is only used in the legacy screening. The
routine would not necessarily find the correct set of extermal energies,
conduction min/max & valence min/max, as it was only searching spin=1
2. Major fix for screen_chi0. The energy array was being index wrong to build
up the Green's function (though depending on compiler this might have worked
out ok). More importantly, there seemed to be a factor of 2 off. This has been
fixed, but there is a bit of asymmetry with respect to builder.f90 which
should be tracked down.